+++ akshare/stock_fundamental/stock_profit_forecast.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2021/5/4 15:12
+Date: 2023/2/9 17:50
 Desc: 东方财富网-数据中心-研究报告-盈利预测
-http://data.eastmoney.com/report/profitforecast.jshtml
+https://data.eastmoney.com/report/profitforecast.jshtml
 """
 import pandas as pd
 import requests
@@ -13,7 +13,7 @@ from tqdm import tqdm
 def stock_profit_forecast():
     """
     东方财富网-数据中心-研究报告-盈利预测
-    http://data.eastmoney.com/report/profitforecast.jshtml
+    https://data.eastmoney.com/report/profitforecast.jshtml
     :return: 盈利预测
     :rtype: pandas.DataFrame
     """
@@ -49,10 +49,11 @@ def stock_profit_forecast():
 
     big_df.reset_index(inplace=True)
     big_df["index"] = big_df.index + 1
-    year1 = list(set(big_df['YEAR1']))[-1]
-    year2 = list(set(big_df['YEAR2']))[-1]
-    year3 = list(set(big_df['YEAR3']))[-1]
-    year4 = list(set(big_df['YEAR4']))[0]
+
+    year1 = str(big_df['YEAR1'].mode().values[0])
+    year2 = str(big_df['YEAR2'].mode().values[0])
+    year3 = str(big_df['YEAR3'].mode().values[0])
+    year4 = str(big_df['YEAR4'].mode().values[0])
     big_df.columns = [
         "序号",
         "-",
@@ -105,16 +106,21 @@ def stock_profit_forecast():
             f"{year4}预测每股收益",
         ]
     ]
-    big_df['研报数'] = pd.to_numeric(big_df['研报数'])
-    big_df['机构投资评级(近六个月)-买入'] = pd.to_numeric(big_df['机构投资评级(近六个月)-买入'])
-    big_df['机构投资评级(近六个月)-增持'] = pd.to_numeric(big_df['机构投资评级(近六个月)-增持'])
-    big_df['机构投资评级(近六个月)-中性'] = pd.to_numeric(big_df['机构投资评级(近六个月)-中性'])
-    big_df['机构投资评级(近六个月)-减持'] = pd.to_numeric(big_df['机构投资评级(近六个月)-减持'])
-    big_df['机构投资评级(近六个月)-卖出'] = pd.to_numeric(big_df['机构投资评级(近六个月)-卖出'])
-    big_df[f"{year1}预测每股收益"] = pd.to_numeric(big_df[f"{year1}预测每股收益"])
-    big_df[f"{year2}预测每股收益"] = pd.to_numeric(big_df[f"{year2}预测每股收益"])
-    big_df[f"{year3}预测每股收益"] = pd.to_numeric(big_df[f"{year3}预测每股收益"])
-    big_df[f"{year4}预测每股收益"] = pd.to_numeric(big_df[f"{year4}预测每股收益"])
+    big_df['机构投资评级(近六个月)-买入'].fillna(0, inplace=True)
+    big_df['机构投资评级(近六个月)-增持'].fillna(0, inplace=True)
+    big_df['机构投资评级(近六个月)-中性'].fillna(0, inplace=True)
+    big_df['机构投资评级(近六个月)-减持'].fillna(0, inplace=True)
+    big_df['机构投资评级(近六个月)-卖出'].fillna(0, inplace=True)
+    big_df['研报数'] = pd.to_numeric(big_df['研报数'], errors="coerce")
+    big_df['机构投资评级(近六个月)-买入'] = pd.to_numeric(big_df['机构投资评级(近六个月)-买入'], errors="coerce")
+    big_df['机构投资评级(近六个月)-增持'] = pd.to_numeric(big_df['机构投资评级(近六个月)-增持'], errors="coerce")
+    big_df['机构投资评级(近六个月)-中性'] = pd.to_numeric(big_df['机构投资评级(近六个月)-中性'], errors="coerce")
+    big_df['机构投资评级(近六个月)-减持'] = pd.to_numeric(big_df['机构投资评级(近六个月)-减持'], errors="coerce")
+    big_df['机构投资评级(近六个月)-卖出'] = pd.to_numeric(big_df['机构投资评级(近六个月)-卖出'], errors="coerce")
+    big_df[f"{year1}预测每股收益"] = pd.to_numeric(big_df[f"{year1}预测每股收益"], errors="coerce")
+    big_df[f"{year2}预测每股收益"] = pd.to_numeric(big_df[f"{year2}预测每股收益"], errors="coerce")
+    big_df[f"{year3}预测每股收益"] = pd.to_numeric(big_df[f"{year3}预测每股收益"], errors="coerce")
+    big_df[f"{year4}预测每股收益"] = pd.to_numeric(big_df[f"{year4}预测每股收益"], errors="coerce")
     return big_df