+++ docs/changelog.md
@@ -49,6 +49,10 @@
 
 ## 更新说明
 
+1.8.78 fix: fix stock_profit_forecast interface
+
+    1. 修复 stock_profit_forecast 接口
+
 1.8.77 fix: fix stock_analyst_rank_em interface
 
     1. 修复 stock_analyst_rank_em 接口
@@ -1857,6 +1861,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.78 fix: fix stock_profit_forecast interface
 
 1.8.77 fix: fix stock_analyst_rank_em interface